requirements.txt
olefile==0.44
Pillow==4.1.1
selenium==3.4.3
